pyproject.toml
[project]
name = "fashion-rag"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "numpy",
    "pandas",
    "Pillow",
    "tqdm",
    "universal-pathlib",
    "gcsfs",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "db-dtypes",
    "scikit-learn>=1.8.0",
    "matplotlib>=3.10.8",
]

[project.optional-dependencies]
ml = [
    "torch",
    "transformers",
]
app = [
    "fashion-rag[ml]",
    "streamlit",
]
vertex = [
    "kfp",
    "google-cloud-aiplatform",
]
eval = [
    "fashion-rag[ml]",
    "umap-learn",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fashion_rag"]
